01-task-manager-postgresql/requirements.txt
psycopg2-binary==2.9.6
Faker==18.9.0
